+++ Keyboards/KeyboardsBase/ScribeFunctionality/CommandBar.swift
@@ -56,7 +56,6 @@ class CommandBar: UILabel {
   }
 
   var shadow: UIButton!
-  var blend: UILabel!
 
   // MARK: - Internal methods
 
@@ -64,9 +63,6 @@ class CommandBar: UILabel {
   func set() {
     addInfoButton()
     backgroundColor = commandBarColor
-    blend.backgroundColor = commandBarColor
-    layer.borderColor = commandBarBorderColor
-    layer.borderWidth = 1.0
     textAlignment = NSTextAlignment.left
     if DeviceType.isPhone {
       font = .systemFont(ofSize: frame.height * 0.4725)
@@ -129,7 +125,6 @@ class CommandBar: UILabel {
     layer.borderColor = UIColor.clear.cgColor
     text = ""
     shadow.backgroundColor = UIColor.clear
-    blend.backgroundColor = UIColor.clear
   }
 
   // Removes the placeholder text for a command and replaces it with just the prompt and cursor.